Clean frontend code
- Move all the logic to a separated file following the atomic design pattern
- Add Performace copy page into the main page
