+++ frontend/src/components/Pages/index.ts
@@ -1,0 +1,1 @@
+export { PerformanceCopy } from "./Performance";
