+++ frontend/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./Atoms";
 export * from "./Molecules";
+export * from "./Pages";
